+++ R/findNextCATItem.R
@@ -1,7 +1,7 @@
 #' Find next CAT item
 #'
 #' A function that returns the next item in the computerized adaptive, optimal assembly, or shadow test.
-#' For direction manupulation of the internal objects this function should be used in conjunction
+#' For direction manipulation of the internal objects this function should be used in conjunction
 #' with the \code{\link{updateDesign}} and \code{customNextItem}.
 #' Finally, the raw input forms can be used when a \code{customNextItem} function has been
 #' defined in \code{\link{mirtCAT}}.
@@ -87,7 +87,7 @@
 #'                   design = list(customNextItem=customNextItem))
 #'                 
 #' -----------------------------------------------------------
-#' # direct manupulation of internal objects
+#' # direct manipulation of internal objects
 #' CATdesign <- mirtCAT(df, mod, criteria = 'MI', design_elements = TRUE)
 #'
 #' # returns number 1 in this case, since that's the starting item